refactored a bit of code to split some struct and load both config.toml and config.json
